Rename send_mail_mandrill.py to senders.py.
